Rename `memo_app.js` to `app.js`
