+++ 04.class/memo.js
@@ -1,7 +1,7 @@
 #!/usr/bin/env node
 
 import minimist from "minimist";
-import { App } from "./lib/memo_app.js";
+import { App } from "./lib/app.js";
 
 const params = minimist(process.argv.slice(2));
 const app = new App();
